+++ RELEASE_NOTES.md
@@ -1,6 +1,6 @@
 
 ### Version 1.1.7
-- Add a file name check to void user uploading files with name starting with space or period
+- Add a file name check to void user uploading files with name starting with space
 
 ### Version 1.1.3
 - Add a add-acl-concierge endpoint